phase6-aws-guardduty/requirements.txt
boto3>=1.34.0
moto[ec2,iam,s3,ssm,guardduty,cloudtrail,sns]>=5.0.0
pytest>=7.4.0
pytest-mock>=3.12.0
python-dotenv>=1.0.0
requests>=2.31.0
